requirements.txt
psycopg2-binary
pandas
python-dotenv
google-cloud-storage
fastapi
uvicorn
openpyxl
requests